Making react optional - users of the npm package would do:
import ReactGA from 'react-ga/noreact'

(cherry picked from commit 245f1a64d234431d567c412ba8f8d02532a28ddb)
