+++ noreact.js
@@ -1,0 +1,2 @@
+var index = require('./dist/react-ga-noreact');
+module.exports = index.default;
